deduplicate file upload
